+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/data/database/history/HistoryForDatabase.java
@@ -136,6 +136,11 @@ public class HistoryForDatabase implements History {
     }
 
     @Override
+    public Line addWithSameHeaderPrefix(Line line) {
+        throw notImplementedYet();
+    }
+
+    @Override
     public void remove(int indexes) {
         throw notImplementedYet();
     }