+++ projects/net.splitcells.gel.core/src/main/java/net/splitcells/gel/solution/SolutionI.java
@@ -173,6 +173,11 @@ public class SolutionI implements Solution {
         return this.problem.add(line);
     }
 
+    @Override
+    public Line addWithSameHeaderPrefix(Line line) {
+        return problem.addWithSameHeaderPrefix(line);
+    }
+
     @SuppressWarnings("all")
     public void remove(final int lineIndex) {
         this.problem.remove(lineIndex);